+++ src/engine/keyinput.c
@@ -12,7 +12,6 @@
 
 #include "engine.h"
 #include "renderer.h"
-#include <stdlib.h>
 
 t_inputhandler	inputhandler__init(void)
 {
@@ -26,7 +25,7 @@ bool	inputhandler__is_action(t_inputhandler *this)
 	return (this->is_up_pressed || this->is_down_pressed
 		|| this->is_left_pressed || this->is_right_pressed
 		|| this->is_left_rotate_pressed || this->is_right_rotate_pressed
-		|| this->is_toggle_door || is_mouse_motion(this) );
+		|| this->is_toggle_door || is_mouse_motion(this));
 }
 
 static void	inputhandler__key_action(
