+++ src/engine/movement.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   movement.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tkim <tkim@student.42seoul.kr>             +#+  +:+       +#+        */
+/*   By: tkim <tkim@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/04 10:13:06 by youkim            #+#    #+#             */
-/*   Updated: 2022/06/10 10:51:03 by tkim             ###   ########.fr       */
+/*   Updated: 2022/06/10 13:32:06 by tkim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,32 +47,14 @@ void	engine__rotate_player(t_engine *e, double angle)
 {
 	camera__rotate(&e->camera, angle * e->rotspeed);
 }
+
 t_mousemotion	is_mouse_motion(const t_inputhandler *this)
 {
 	if (this->mouse_motion_size / 10 == 0)
-		return MOUSEMOTION_NONE;
+		return (MOUSEMOTION_NONE);
 	if (this->x < WIDTH / 3)
-		return MOUSEMOTION_LEFT;
+		return (MOUSEMOTION_LEFT);
 	else if (this->x > WIDTH * 2 / 3)
-		return MOUSEMOTION_RIGHT;
-	return MOUSEMOTION_NONE;
-}
-
-void	engine__move_player(t_engine *e)
-{
-	const t_inputhandler	*ih = &e->inputhandler;
-
-	if (ih->is_up_pressed)
-		engine__try_move_player_straight(e, POSITIVE);
-	if (ih->is_down_pressed)
-		engine__try_move_player_straight(e, NEGATIVE);
-	if (ih->is_left_pressed)
-		engine__try_strafe_player(e, NEGATIVE);
-	if (ih->is_right_pressed)
-		engine__try_strafe_player(e, POSITIVE);
-	if (ih->is_right_rotate_pressed || is_mouse_motion(ih) == MOUSEMOTION_RIGHT)
-		engine__rotate_player(e, POSITIVE);
-	if (ih->is_left_rotate_pressed || is_mouse_motion(ih) == MOUSEMOTION_LEFT)
-		engine__rotate_player(e, NEGATIVE);
-
+		return (MOUSEMOTION_RIGHT);
+	return (MOUSEMOTION_NONE);
 }
