+++ src/parser/check__map.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check__map.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: youkim    <youkim@student.42seoul.kr>      +#+  +:+       +#+        */
+/*   By: tkim <tkim@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/04 10:13:06 by youkim            #+#    #+#             */
-/*   Updated: 2022/05/04 10:13:06 by youkim           ###   ########.fr       */
+/*   Updated: 2022/06/10 13:28:48 by tkim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,26 +48,4 @@ void	mapformat__assert_valid(
 	if (!(func(this)))
 		std__panic__value__char(
 			"raw_map_arr__check_valid: invalid tile", this);
-}
-
-void	raw_map_arr__check_valid(t_string_arr this)
-{
-	t_ivec						it;
-	t_mapfmt					tile;
-	t_mapformat__validator_f	func;
-
-	if (BONUS)
-		func = mapformat__is_valid_bonus;
-	else
-		func = mapformat__is_valid;
-	it.y = -1;
-	while (this[++it.y])
-	{
-		it.x = -1;
-		while (this[it.y][++it.x])
-		{
-			tile = this[it.y][it.x];
-			mapformat__assert_valid(tile, func);
-		}
-	}
 }
